use "read" stats (rate by days) instead of "finished"

+++ lo3common/src/main/scala/models/api/TagDataPoint.scala
@@ -3,7 +3,7 @@ package models.api
 
 import zio.schema.{DeriveSchema, Schema}
 
-case class TagDataPoint (label: String, value: Int)
+case class TagDataPoint (label: String, value: Long)
 object  TagDataPoint:
   implicit val schema: Schema[TagDataPoint] = DeriveSchema.gen
